CRUD Disciplines in progress

+++ Blazor/MyAcademy/Models/Discipline.cs
@@ -7,11 +7,11 @@ namespace MyAcademy.Models;
 public partial class Discipline
 {
     [Key]
-    public int discipline_id { get; set; }
+    public short discipline_id { get; set; }
 
     public string? discipline_name { get; set; }
 
-    public int number_of_lessons { get; set; }
+    public byte number_of_lessons { get; set; }
 
     public virtual ICollection<Exam> Exams { get; set; } = new List<Exam>();
